vue.js and new restful services added

+++ farma-app/backend/src/main/java/com/it529/teamgy/pharmacyapp/repository/CountryRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 
 @Repository("countryRepository")
 public interface CountryRepository extends JpaRepository<Country, Integer> {
-    Country findById(int id);
+    Country findById(long id);
 
     @Override
     Page<Country> findAll(Pageable pageable);
